+++ internal/resolvers/betterttv/emote_resolver.go
@@ -16,18 +16,18 @@ type EmoteResolver struct {
 	emoteCache cache.Cache
 }
 
-func (r *EmoteResolver) Check(ctx context.Context, url *url.URL) bool {
+func (r *EmoteResolver) Check(ctx context.Context, url *url.URL) (context.Context, bool) {
 	// Ensure that the domain is either betterttv.com or www.betterttv as defined in the domains map in initialize.go
 	if match, _ := resolver.MatchesHosts(url, domains); !match {
-		return false
+		return ctx, false
 	}
 
 	// Ensure that the path of the url matches the emote path regex as defined in initialize.go
 	if !emotePathRegex.MatchString(url.Path) {
-		return false
+		return ctx, false
 	}
 
-	return true
+	return ctx, true
 }
 
 func (r *EmoteResolver) Run(ctx context.Context, url *url.URL, req *http.Request) ([]byte, error) {