+++ internal/resolvers/betterttv/emote_resolver_test.go
@@ -69,7 +69,7 @@ func TestEmoteResolver(t *testing.T) {
 
 		for _, test := range tests {
 			c.Run(test.label, func(c *qt.C) {
-				output := resolver.Check(ctx, test.input)
+				_, output := resolver.Check(ctx, test.input)
 				c.Assert(output, qt.Equals, test.expected)
 			})
 		}